jwt token, usermodel and private routes added

+++ controllers/contactController.js
@@ -3,45 +3,46 @@ const Contact=require("../models/contactModel");
 
 //@desc Get all contacts
 //@route GET /api/contacts
-//@access public
+//@access private
 
 const getContacts=expressAsyncHandler(async(req,res)=>{
-    const contacts=await Contact.find();
+    const contacts=await Contact.find({user_id:req.user.id});
     res.status(200).json(contacts);
 });
 
 //@desc Get contact by id
 //@route GET /api/contacts
-//@access public
+//@access private
 
-const getContact=async(req,res)=>{
+const getContact=expressAsyncHandler(async(req,res)=>{
     const contact=await Contact.findById(req.params.id)
     if(!contact){
         res.status(404);
         throw new Error("Contact not Found");
     }
     res.status(200).json(contact);
-}
+});
 
 //@desc create contact 
 //@route POST /api/contacts
-//@access public
+//@access private
 
-const createContact=async(req,res)=>{
+const createContact=expressAsyncHandler(async(req,res)=>{
     const {name, email, phone}=req.body;
     if(!name || !email || !phone){
         res.status(400);
         throw new Error('All fields are mandatory!...');
     }
     const contact=await Contact.create({
+        user_id:req.user.id,
         name,email,phone
     })
     res.status(200).json(contact);
-}
+});
 
 //@desc update contact by id
 //@route PUT /api/contacts
-//@access public
+//@access private
 
 const updateContact=expressAsyncHandler(async(req,res)=>{
     const contact=await Contact.findById(req.params.id)
@@ -49,6 +50,10 @@ const updateContact=expressAsyncHandler(async(req,res)=>{
         res.status(404);
         throw new Error("Contact not Found");
     } 
+    if(contact.user_id.toString()!==req.user.id){
+        res.status(403);
+        throw new Error("user don't have permission to update other user contacts")
+    }
 
     const updatedContact=await Contact.findByIdAndUpdate(
         req.params.id,
@@ -70,6 +75,12 @@ const deleteContact=expressAsyncHandler(async(req,res)=>{
         res.status(404);
         throw new Error("Contact not Found");
     } 
+
+    if(contact.user_id.toString()!==req.user.id){
+        res.status(403);
+        throw new Error("user don't have permission to update other user contacts")
+    }
+    
     await Contact.deleteOne({_id: req.params.id});
     res.status(200).json(contact);
 });
